==> Auxiliary/Step3_MaskImage.ipynb <==
import numpy as np
from PIL import Image
import seaborn as sns;
import matplotlib.pyplot as plt
import pandas as pd

from skimage import measure
from skimage import filters
from skimage.filters import threshold_otsu

XC = XC[:2000]


X.shape

X,Y = np.meshgrid(np.linspace(0,500,3000),np.linspace(0,500,10))

def GetIndexTest(X,Y,XC):

    n_x,n_y     = X.shape[1],X.shape[0];
    dx,dy       = X[1,1]-X[0,0],Y[1,1]-Y[0,0];
    x_min,y_min = X[0,0],Y[0,0];

    x_index = (np.round((XC[:,0]-x_min)/dx)).astype(int)
    y_index = (np.round((XC[:,1]-y_min)/dy)).astype(int)
    
    x_index = np.minimum(n_x-1,np.maximum(0,x_index));
    y_index = np.minimum(n_y-1,np.maximum(0,y_index));    
    
#    print(x_index)
 #   print(y_index)  
    return y_index*n_x+x_index

def TestIndexing(X,Y,XC):
    idx = GetIndexTest(X,Y,XC);
    X_flat = X.flatten()
    Y_flat = Y.flatten()
    dx = (X[1,1]-X[0,0])/2;
    dy = (Y[1,1]-Y[0,0])/2;    
    print(XC[:,0]-X_flat[idx]<dx);
    print(XC[:,1]-Y_flat[idx]<dy);
    print('Max distance dx/2: '+str(dx))
    print('Max distance dy/2: '+str(dy))

TestIndexing(X,Y,XC[80:120,:])

XC[:3,:]

np.min(Y)

Y[0,0]

X_flat

X.shape[0]

XX,YY = np.meshgrid([1,2,3],[4,5]);

YY.shape

len(np.unique(Y))

#*****

n = 1000;

filename      = "../../ProteinData_ttx_1hr_2/AHA_2_MMStack_Pos0.ome_locs_render_driftcor_filter_render_pix0.02X6f20";
#filename = "../MikeData/EGFR-P1-ATTO655_cell_3_MMImages.ome_locs_render"
#filename = "../../MikeData/EGFR-P1-ATTO655_cell_2_MMImages.ome_locs_render_al_linked1sigma"

XC = np.loadtxt(filename+".txt");

H_mask_signal = np.loadtxt(filename+"_heatmap_mask_signal.txt");
H_mask_noise  = np.loadtxt(filename+"_heatmap_mask_noise.txt");

H = np.loadtxt(filename+"_heatmap_D.txt");
X = np.loadtxt(filename+"_heatmap_X.txt");
Y = np.loadtxt(filename+"_heatmap_Y.txt");

X_flat = X.flatten()
Y_flat = Y.flatten()
H_noise_flat = (H_mask_noise.T).flatten()
H_signal_flat = (H_mask_signal.T).flatten()

H_mask_noise.shape

if(X.shape != (n,n)):
    raise Exception("Wrong dimension");

XC_u = np.unique(X);
dx = np.mean(XC_u[1:]-XC_u[:-1])
x_min = np.min(XC_u);

YC_u = np.unique(Y);
dy = np.mean(YC_u[1:]-YC_u[:-1])
y_min = np.min(YC_u);

dy

print(X[0,1]-X[0,0])
print(Y[1,1]-Y[0,0])

def GetIndex(xc):
    x_index = (np.round((xc[:,0]-x_min)/dx)).astype(int)
    y_index = (np.round((xc[:,1]-y_min)/dy)).astype(int)
    
    x_index = np.minimum(n-1,np.maximum(0,x_index));
    y_index = np.minimum(n-1,np.maximum(0,y_index));    
    
#    print(x_index)
 #   print(y_index)  
    return y_index*n+x_index

def TestIndex(XC):
     index =  GetIndex(XC)
     print((X_flat[index] - XC[:,0]<dx));
     print((Y_flat[index] - XC[:,1]<dy));

mask_noise  = H_noise_flat[GetIndex(XC)];
mask_signal = H_signal_flat[GetIndex(XC)];

XC_signal = XC[mask_signal==True];
XC_noise  = XC[mask_noise==True];

np.savetxt(filename+"_X_signal.txt",XC_signal,fmt="%f");
np.savetxt(filename+"_X_noise.txt",XC_noise,fmt="%f");

H_signal, xedges, yedges = np.histogram2d(XC_signal[:,0],XC_signal[:,1],bins=1000);
H_noise, xedges, yedges = np.histogram2d(XC_noise[:,0],XC_noise[:,1],bins=1000);

plt.imshow(H_signal>0)
plt.savefig(filename+'_Image_Signal.pdf')

plt.imshow(H_noise>0)
plt.savefig(filename+'_Image_Noise.pdf')

plt.imshow(H>0)
plt.savefig(filename+'_Image_all.pdf')

plt.imshow((H_noise==0)*(H_signal==0))
plt.savefig(filename+'_Image_all.pdf')
